Added assumptions to readme added more unit tests calculated quiet streak

+++ RssSummary.cs
@@ -7,7 +7,7 @@ namespace QuietRssNuget
     //TODO: Ask on longevity of this code
     public class RssSummary
     {
-        public RssSummary(string company, string rssUri, DateTimeOffset lastUpdate, int quietStreak, string title)
+        public RssSummary(string company, string rssUri, DateTimeOffset lastUpdate, double quietStreak, string title)
         {
             this.Company = company;
             this.RssUri = rssUri;
@@ -20,7 +20,7 @@ namespace QuietRssNuget
         public string Company {get;set;}
         public string Title { get; set; }
         public string RssUri { get; set; }
-        public int QuietStreak {get;set;}
+        public double QuietStreak {get;set;}
         public DateTime LastPostDateUTC { get; set; }
 
     }
